Reworking the signal r samples (WIP)
