--- projects/signalr-1/Client/global.json
@@ -1,5 +1,0 @@
-{
-  "sdk": {
-    "version": "2.0.0"
-  }
-}
