--- projects/signalr-1/Server/Startup.cs
@@ -1,36 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace signalr
-{
-    public class Startup
-    {
-        public void ConfigureServices(
-            IServiceCollection services)
-        {
-            services.AddCors(options =>
-            {
-                options.AddPolicy("all",
-                    policy => policy.AllowAnyOrigin()
-                                    .AllowAnyHeader()
-                                    .AllowAnyMethod());
-            });
-
-            services.AddSignalR();
-        }
-
-        public void Configure(
-            IApplicationBuilder app, 
-            IHostingEnvironment env)
-        {
-            app.UseDeveloperExceptionPage();
-            app.UseCors("all");
-
-            app.UseSignalR(routes =>
-            {
-                routes.MapHub<ChatHub>("chatHub");
-            });
-        }
-    }
-}
