--- projects/signalr-1/Server/global.json
@@ -1,5 +1,0 @@
-{
-  "sdk": {
-    "version": "2.0.0"
-  }
-}
