+++ projects/signalr/global.json
@@ -1,5 +1,5 @@
 {
   "sdk": {
-    "version": "2.1.300-preview1-008174"
+    "version": "2.1.300-preview2-008530"
   }
 }
