+++ projects/signalr/signalr-1/.gitignore
