+++ projects/signalr/signalr-1/Client/.vscode/tasks.json
@@ -1,0 +1,15 @@
+{
+    "version": "2.0.0",
+    "tasks": [
+        {
+            "label": "build",
+            "command": "dotnet",
+            "type": "process",
+            "args": [
+                "build",
+                "${workspaceFolder}/Client.csproj"
+            ],
+            "problemMatcher": "$msCompile"
+        }
+    ]
+}
