+++ projects/signalr/signalr-1/Client/package.json
@@ -3,8 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "dependencies": {
-    "@aspnet/signalr-client": "^1.0.0-alpha1-final",
-    "jquery": "^3.2.1"
+    "@aspnet/signalr": "1.0.0-preview2-final"
   },
   "devDependencies": {},
   "license": "MIT"
