+++ projects/signalr/signalr-1/Client/src/Controllers/HomeController.cs
