+++ projects/signalr/signalr-1/Client/src/Views/Shared/_Layout.cshtml
@@ -4,8 +4,8 @@
     <meta name="viewport" content="width=device-width" />
     <title>ASP.NET Core SignalR</title>
 
-    <script src="./node_modules/@@aspnet/signalr-client/dist/browser/signalr-client-1.0.0-alpha2-final.js"></script>
-    <script src="./node_modules/jquery/dist/jquery.js"></script>
+    <script src="/node_modules/@@aspnet/signalr/dist/browser/signalr.js"></script>
+    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha256-3edrmyuQ0w65f8gfBsqowzjJe2iM6n0nKciPUp8y+7E=" crossorigin="anonymous"></script>
 </head>
 <body>
     <div>
