+++ projects/signalr/signalr-1/Client/src/Views/_ViewStart.cshtml
