+++ projects/signalr/signalr-2/global.json
