+++ projects/signalr/signalr-2/signalr-2.csproj
