+++ projects/signalr/signalr-2/src/Startup.cs
